feat: add meta param to the portStream message. continueInterval can be set using the meta param
The meta parameter can be used to pass a variety of additional parameters, with continueInterval
being just one of them. portStream sets the send of the continue message by adding
cont(continueInterval) to the meta parameter of the Open/Req message to direct the peer. Thus the
portStream does not need to specify a continueInterval at creation.

BREAKING CHANGE: The new version of pull-mesh cannot communicate with the old version of pull-mesh
due to a change in message format.